Extract subcommands to separate modules
