+++ wafflectl/src/commands/mod.rs
@@ -1,0 +1,9 @@
+mod add;
+mod delete;
+mod get;
+mod run;
+
+pub use add::AddSubcommand;
+pub use delete::DeleteSubcommand;
+pub use get::GetSubcommand;
+pub use run::RunSubcommand;
